% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper-mutate.R
\name{hyper_mutate.data.frame}
\alias{hyper_mutate.data.frame}
\title{Hyper Mutate implementation for Data Frames}
\usage{
\method{hyper_mutate}{data.frame}(.data, ...)
}
\arguments{
\item{...}{}
}
\description{
Preferred over tibbles as tibble transformations can be slower than
data.frame. Can be compared using:
}
