% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper-mutate.R
\name{hyper_mutate}
\alias{hyper_mutate}
\alias{hyper_mutate.data.frame}
\alias{hyper_mutate.vector}
\title{An efficient alternative to dplyr's mutate function}
\usage{
hyper_mutate(.data, ...)

\method{hyper_mutate}{data.frame}(.data, ...)

\method{hyper_mutate}{vector}(.data, ...)
}
\arguments{
\item{...}{}
}
\description{
`hyper_mutate` similiar to `dplyr::mutate` is a lightweight and highly
efficient data manipulation function,that addresses some of the performance
limitations found in dplyr.
}
