% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper-mutate.R
\name{mutate_class_wrapper}
\alias{mutate_class_wrapper}
\title{Mutate Data Frame R6 Class wrapper}
\usage{
mutate_class_wrapper(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{<[`data-masking`][rlang::args_data_masking]> Name-value pairs.
The name gives the name of the column in the output.}
}
\description{
Class wrapper for `MutateDataFrame` R6 class
}
