% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper-mutate.R
\name{MutateDataFrame}
\alias{MutateDataFrame}
\title{Mutate Data Frame R6 Class implementation}
\description{
R6 class for modifying/mutating data frames
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MutateDataFrame-new}{\code{MutateDataFrame$new()}}
\item \href{#method-MutateDataFrame-mutate_class}{\code{MutateDataFrame$mutate_class()}}
\item \href{#method-MutateDataFrame-clone}{\code{MutateDataFrame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MutateDataFrame-new"></a>}}
\if{latex}{\out{\hypertarget{method-MutateDataFrame-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutateDataFrame$new(data_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MutateDataFrame-mutate_class"></a>}}
\if{latex}{\out{\hypertarget{method-MutateDataFrame-mutate_class}{}}}
\subsection{Method \code{mutate_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutateDataFrame$mutate_class(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MutateDataFrame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MutateDataFrame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutateDataFrame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
