% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper-mutate.R
\name{mutate_class_wrapper}
\alias{mutate_class_wrapper}
\title{Mutate Data Frame R6 Class wrapper}
\usage{
mutate_class_wrapper(.data, ...)
}
\arguments{
\item{...}{}
}
\description{
Class wrapper for `MutateDataFrame` R6 class
}
